src/third_party/blink/web_tests/http/tests/contacts/select-function.html

<!doctype html>
<meta charset="utf-8">
<title>Contact API: Behaviour of the select() function</title>
<script src="/gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="/gen/third_party/blink/public/mojom/contacts/contacts_manager.mojom.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>
<script src="resources/mock_contacts_manager.js"></script>
<script>
'use strict';

// Verifies that |func|, when invoked, throws a TypeError exception.
async function expectTypeError(func) {
  try {
    await func();
  } catch (e) {
    assert_equals(e.name, 'TypeError');
    return;
  }

  assert_unreached('expected a TypeError, but none was thrown');
}

test(() => {
  // Exposure of the interface and method.
  assert_own_property(window, 'ContactsManager');
  assert_own_property(ContactsManager.prototype, 'select');

  // Exposure of the instance.
  assert_idl_attribute(navigator, 'contacts');
  assert_idl_attribute(navigator.contacts, 'select');

}, 'The Contact API is exposed on the Window context');

promise_test(async () => {
  await expectTypeError(() =>
      navigator.contacts.select(['name']));

}, 'The Contact API requires a user gesture')

promise_test(async () => {
  triggerUserGesture();

  // At least one property must be provided.
  await expectTypeError(() => navigator.contacts.select());
  await expectTypeError(() => navigator.contacts.select([]));

  // Per WebIDL parsing, no invalid values may be provided.
  await expectTypeError(() =>
      navigator.contacts.select(['']));
  await expectTypeError(() =>
      navigator.contacts.select(['foo']));
  await expectTypeError(() =>
      navigator.contacts.select(['name', 'photo']));

}, 'The Contact API requires valid properties to be provided');

promise_test(async () => {
  triggerUserGesture();

  // Returns a NULL result, indicating that no results are available.
  mockContactsManager.setSelectCallback(async (options) => {
    return { contacts: null };
  });

  await expectTypeError(() => navigator.contacts.select(['name']));

}, 'The Contact API can fail when the selector cannot be opened');

promise_test(async () => {
  triggerUserGesture();

  let storedOptions = null;

  // Stores the |options| in |storedOptions| for inspection after the call
  // completes. Deliberately fails the call with no data.
  mockContactsManager.setSelectCallback(async (options) => {
    storedOptions = options;

    return { contacts: null };
  });

  await expectTypeError(() => {
    return navigator.contacts.select(['name', 'email'], { multiple: true });
  });

  assert_not_equals(storedOptions, null);
  assert_true(storedOptions.multiple);
  assert_true(storedOptions.includeNames);
  assert_true(storedOptions.includeEmails);
  assert_false(storedOptions.includeTel);

}, 'The Contact API correctly interprets the JavaScript arguments');

promise_test(async () => {
  triggerUserGesture();

  // Returns two contacts with all information available.
  mockContactsManager.setSelectCallback(async (options) => {
    return {
      contacts: [
        { name: ['Finnur'], email: ['finnur@chromium.org'], tel: ['000-0000'] },
        { name: ['Peter', 'Test'], email: ['peter@chromium.org'], tel: [] },
      ]
    };
  });

  const results = await navigator.contacts.select(['name', 'email', 'tel'], { multiple: true });

  assert_equals(results.length, 2);

  {
    const finnur = results[0];

    assert_own_property(finnur, 'name');
    assert_own_property(finnur, 'email');
    assert_own_property(finnur, 'tel');

    assert_array_equals(finnur.name, ['Finnur']);
    assert_array_equals(finnur.email, ['finnur@chromium.org']);
    assert_array_equals(finnur.tel, ['000-0000']);
  }

  {
    const peter = results[1];

    assert_own_property(peter, 'name');
    assert_own_property(peter, 'email');
    assert_own_property(peter, 'tel');

    assert_array_equals(peter.name, ['Peter', 'Test']);
    assert_array_equals(peter.email, ['peter@chromium.org']);
    assert_array_equals(peter.tel, []);
  }

}, 'The Contact API correctly returns ContactInfo entries');

promise_test(async () => {
  triggerUserGesture();

  // Returns partial information since no e-mail addresses are requested.
  mockContactsManager.setSelectCallback(async (options) => {
    return {
      contacts: [{ name: ['Tim'], email: null, tel: null }]
    };
  });

  const results = await navigator.contacts.select(['name']);

  assert_equals(results.length, 1);

  {
    const tim = results[0];

    assert_own_property(tim, 'name');
    assert_false('email' in tim);
    assert_false('tel' in tim);

    assert_array_equals(tim.name, ['Tim']);
    assert_equals(tim.email, undefined);
    assert_equals(tim.tel, undefined);
  }
}, 'The Contact API does not include fields that were not requested');

promise_test(async () => {
  triggerUserGesture();

  // Returns partial information since no e-mail addresses are requested.
  mockContactsManager.setSelectCallback(async (options) => {
    return {
      contacts: [{ name: ['Tim'], email: null, tel: null }]
    };
  });

  // First request should work.
  const promise1 = new Promise((resolve, reject) => {
    navigator.contacts.select(['name']).then(resolve)
                                       .catch(e => reject(e.message));
  });

  // Second request should fail (since the first one didn't complete yet).
  const promise2 = new Promise((resolve, reject) => {
    navigator.contacts.select(['name']).then(contacts => reject('This was supposed to fail'))
                                       .catch(e => resolve(e.message));
  });

  const results = await Promise.all([promise1, promise2]);
  const contacts = results[0];
  assert_equals(contacts.length, 1);
  const contact = contacts[0];
  assert_equals(contact.name[0], 'Tim');
  assert_equals(results[1], 'Contacts Picker is already in use.');
}, 'The Contact API cannot be used again until the first operation is complete.');

</script>
